Restructure DropdownModule component + friends
