--- client/components/dropdownModule/Assessment.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import Exercise from './Exercise'
-
-const Assessment = () => {
-  return (
-    <React.Fragment>
-      <h3>[Placeholder - Assessments] </h3>
-      <Exercise />
-    </React.Fragment>
-  )
-}
-
-export default Assessment
