--- client/components/dropdownModule/EvidenceUrl.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const EvidenceUrl = () => {
-  return (
-    <React.Fragment>
-      <h3>[Placeholder Completion Evidence - Github repo URL] - </h3>
-      <input type="text" name="evidence" placeholder="inside link to your github repo here"></input>
-    </React.Fragment>
-  )
-}
-
-export default EvidenceUrl
