--- client/components/dropdownModule/Exercise.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import EvidenceUrl from './EvidenceUrl'
-import Status from './Status'
-const Exercise = () => {
-  return (
-    <React.Fragment>
-      <h3>[Placeholder Exercises] </h3>
-      <EvidenceUrl/>
-      <Status/>
-    </React.Fragment>
-  )
-}
-
-export default Exercise
