--- client/components/dropdownModule/Modules.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import Assessment from './Assessment'
-
-const Modules = () => {
-  return (
-    <React.Fragment>
-      <h2>[Placeholder - Modules] </h2>
-      <Assessment />
-    </React.Fragment>
-  )
-}
-
-export default Modules
